imdb/requirements.txt
Django==1.11.16
djangorestframework==3.9.0
IMDbPY==6.6
lxml==4.2.5
Pygments==2.3.0
pytz==2018.7
SQLAlchemy==1.2.14
gunicorn
